move indicator to meta data section
